+++ perl/LinkedSpec/RuleIR/EmitContext.pm
@@ -398,6 +398,11 @@ sub _lower_push_value_statement {
  return _call_actionir_owner_with_deps('method_lowering', '_lower_push_value_statement', @args)
 }
 
+sub _lower_push_nonempty_statement {
+ my @args = @_;
+ return _call_actionir_owner_with_deps('method_lowering', '_lower_push_nonempty_statement', @args)
+}
+
 sub _lower_regex_subst_statement {
  my @args = @_;
  return _call_actionir_owner_with_deps('method_lowering', '_lower_regex_subst_statement', @args)